requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
motor
pydantic
